+++ CHANGELOG.md
@@ -45,6 +45,8 @@ IMPROVEMENTS:
   "Packer Builder" so that they are easily recognizable. [GH-642]
 * builder/amazon/all: Copying AMIs to multiple regions now happens
   in parallel. [GH-495]
+* builder/docker: A "run\_command" can be specified, configuring how
+  the container is started. [GH-648]
 * builder/openstack: In debug mode, the generated SSH keypair is saved
   so you can SSH into the machine. [GH-746]
 * builder/qemu: Floppy files are supported. [GH-686]